reorganized codes 	modified:   app.py 	modified:   common.py 	deleted:    html/header.html 	new file:   html/header.tpl 	deleted:    html/header_loggedin.tpl 	new file:   html/menu_loggedin.tpl 	new file:   html/menu_loggedout.html 	new file:   html/setting.tpl 	deleted:    index.py 	modified:   lib/auth.py 	modified:   lib/middleware.py 	modified:   login.py 	modified:   register.py 	modified:   setting.py

+++ login.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
 from urlparse import parse_qs
-from lib import auth, content
+from lib import auth, content, middleware
 import common
 
-def get_handler(environ, start_response, message=''):
-    status = '200 OK'
-    response_headers = [('Content-type', 'text/html')]
-    
+def get_handler(environ, start_response, message=''):    
     template = content.get_template("main.tpl")
-    header = content.get_html("header.html")
+    header = common.header_html(environ)
     body = content.get_template("login.tpl").substitute(message=message)
     html = template.substitute(header=header, body=body)
+
+    status = '200 OK'
+    response_headers = [('Content-type', 'text/html')]
     start_response(status, response_headers)
     return [html]
 
@@ -29,8 +29,11 @@ def post_handler(environ, start_response):
         return get_handler(environ, start_response, "UsernameとPasswordを入力してください。")
 
 def logout(environ, start_response):
+    if not auth.Authenticator.authenticated(environ):
+        return common.redirect_top(environ, start_response)
     tpl = content.get_template("main.tpl")
-    header = content.get_html("header.html")
+    environ.pop('Authenticated')
+    header = common.header_html(environ)
     html = tpl.substitute(header=header, body="<h3>ログアウトしました。</h3>")
     cookie = auth.delete_cookie()
     status = '200 OK'
@@ -48,3 +51,4 @@ def application(environ, start_response):
             return get_handler(environ, start_response)
     elif method == 'POST':
         return post_handler(environ, start_response)
+
